Update new contract type and insert new contract record
