+++ cfg/const.go
@@ -60,6 +60,7 @@ const (
 
 	SMCTypePrefix         = "SMCType:"
 	SMCTypeKRC20          = "KRC20"
+	SMCTypeNormal         = "Normal"
 	SMCTypeKRC721         = "KRC721"
 	SMCTypeValidator      = "Validator"
 	KRCTransferTopic      = "0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef"
