+++ server/info.go
@@ -360,36 +360,6 @@ func (s *infoServer) ProcessTxs(ctx context.Context, block *types.Block, writeTo
 	return nil
 }
 
-func (s *infoServer) ProcessActiveAddress(ctx context.Context, txs []*types.Transaction) error {
-	lgr := s.logger.With(zap.String("method", "ProcessActiveAddress"))
-	// update active addresses
-	startTime := time.Now()
-	addrsMap := filterAddrSet(txs)
-	getBalanceTime := time.Now()
-	addrsList := s.getAddressBalances(ctx, addrsMap)
-	lgr.Debug("GetAddressBalance time", zap.Duration("TotalTime", time.Since(getBalanceTime)))
-
-	updateAddressTime := time.Now()
-	if err := s.dbClient.UpdateAddresses(ctx, addrsList); err != nil {
-		return err
-	}
-	lgr.Debug("UpdateAddressTime", zap.Duration("TotalTime", time.Since(updateAddressTime)))
-	endTime := time.Since(startTime)
-	s.metrics.RecordInsertActiveAddressTime(endTime)
-	s.logger.Info("Total time for update addresses", zap.Duration("TimeConsumed", endTime), zap.String("Avg", s.metrics.GetInsertActiveAddressTime()))
-	startTime = time.Now()
-	totalAddr, totalContractAddr, err := s.dbClient.GetTotalAddresses(ctx)
-	if err != nil {
-		return err
-	}
-	err = s.cacheClient.UpdateTotalHolders(ctx, totalAddr, totalContractAddr)
-	if err != nil {
-		return err
-	}
-	s.logger.Info("Total time for getting active addresses", zap.Duration("TimeConsumed", time.Since(startTime)))
-	return nil
-}
-
 func (s *infoServer) ProcessLogsOfTxs(ctx context.Context, txs []*types.Transaction, blockTime time.Time) error {
 	lgr := s.logger.With(zap.String("method", "processLogsOfTxs"))
 	poolSize := 4
@@ -643,48 +613,6 @@ func (s *infoServer) VerifyBlock(ctx context.Context, blockHeight uint64, networ
 		return true, nil
 	}
 	return false, nil
-}
-
-func filterAddrSet(txs []*types.Transaction) map[string]*types.Address {
-	addrs := make(map[string]*types.Address)
-	for _, tx := range txs {
-		addrs[tx.From] = &types.Address{
-			Address:    tx.From,
-			IsContract: false,
-		}
-		addrs[tx.To] = &types.Address{
-			Address:    tx.To,
-			IsContract: false,
-		}
-		addrs[tx.ContractAddress] = &types.Address{
-			Address:    tx.ContractAddress,
-			IsContract: true,
-		}
-	}
-	delete(addrs, "")
-	delete(addrs, "0x")
-	return addrs
-}
-
-func filterContracts(txs []*types.Transaction) map[string]*types.Address {
-	addrs := make(map[string]*types.Address)
-	for _, tx := range txs {
-		addrs[tx.From] = &types.Address{
-			Address:    tx.From,
-			IsContract: false,
-		}
-		addrs[tx.To] = &types.Address{
-			Address:    tx.To,
-			IsContract: false,
-		}
-		addrs[tx.ContractAddress] = &types.Address{
-			Address:    tx.ContractAddress,
-			IsContract: true,
-		}
-	}
-	delete(addrs, "")
-	delete(addrs, "0x")
-	return addrs
 }
 
 func (s *infoServer) getAddressBalances(ctx context.Context, addrs map[string]*types.Address) []*types.Address {
